mlx_gr00t/requirements.txt
mlx>=0.22.0
mlx-lm>=0.20.0
torch>=2.3.0
torchvision
transformers>=4.51.0
huggingface_hub
numpy
Pillow
einops
